% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation.R
\name{commonEpiIDs}
\alias{commonEpiIDs}
\title{Removes epitopes not analyzed by all datasets being compared}
\usage{
commonEpiIDs(...)
}
\arguments{
\item{...}{at least 2 matrices of z scores with an epitope id column titled id}
}
\value{
the adjusted matrix or vector as a data frame
}
\description{
Discards data for epitopes not analyzed in all datasets entered as arguments
}
