% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{cor.fxn}
\alias{cor.fxn}
\title{Computes Correlation}
\usage{
cor.fxn(toSearch, sampleToMatch)
}
\arguments{
\item{toSearch}{a database matrix of z scores for samples to be compared against sampletoMatch or vector of z scores for a single sample}

\item{sampleToMatch}{a vector of z scores from a sample to compare against database. Must contain same number of rows as toSearch.}
}
\value{
correlation matrix
}
\description{
This function computes the correlation between a sample of interest and another sample of interest or database of samples
}
