% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{epiCount}
\alias{epiCount}
\title{Generates Data Frame that contains number of hits corresponding to each level of a factor in epitopeKey}
\usage{
epiCount(hitDF, key, group.by)
}
\arguments{
\item{hitDF}{a data frame of hits with epitope ids in the first column as generated by z2Hit with keepID=TRUE}

\item{key}{dataframe containing epitope ids and relevant metadata. The first column must be named 'id' and contain eptiope ids}

\item{group.by}{variable in epitope key to count}
}
\value{
data frame
}
\description{
Generates Data Frame that contains number of hits from an input data frame that correspond to each level of a selected
factor in an EpitopeKey
}
