% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation.R
\name{epIdentity}
\alias{epIdentity}
\title{Links a dataframe with epitope ids to epitope metadata}
\usage{
epIdentity(
  dataDF,
  key,
  keepAllData = TRUE,
  keepAllMeta = TRUE,
  keepData = names(dataDF),
  keepMeta = names(key)
)
}
\arguments{
\item{dataDF}{a dataframe of where rows represent epitopes. The first column must be named 'id' and contain eptiope ids}

\item{key}{dataframe containing epitope ids and relevant metadata. The first column must be named 'id' and contain eptiope ids}

\item{keepAllData, KeepAllMeta}{(optional) logical. If False, user may specify columns to exclude in output}

\item{keepData, KeepMeta}{(optional). Specifies which columns to keep if keepAllData or keepAllMeta is FALSE. id must be included. Usage keepData=c("id","var1",...)}
}
\value{
a data frame with epitope meta data and selected individual or database data
}
\description{
Binds columns of data frame with epitope ids to appropriate epitope metadata located in an eptiope
key data frame
}
