% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epitopeKey}
\alias{epitopeKey}
\title{Metadata for epitope ID's}
\format{A data frame with 4 variables:
 \describe{
  \item{id}{id used to identify the epitope recognized}
  \item{Species}{The species of the pathogen from which the epitope originates}
  \item{taxon_family}{family to which that species belongs}
  \item{Protein_name}{protein in which the epitope is found}}}
\usage{
data(epitopeKey)
}
\description{
Defines the species, family, and protein name for each epitope
id in the dataset
}
\keyword{datasets}
