% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation.R
\name{setNA2Val}
\alias{setNA2Val}
\title{Sets NA entries equal to a value}
\usage{
setNA2Val(X, value = 0)
}
\arguments{
\item{X}{a matrix of vector of values}

\item{value}{a numeric value to replace NA's with. Default is 0.}
}
\value{
the adjusted matrix or vector as a data frame
}
\description{
Replaces all NA entries in a vector/matrix with a specified value
}
