% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation.R
\name{setThres2Val}
\alias{setThres2Val}
\title{Set values that are less/greater than or equal to a threshold equal to a value}
\usage{
setThres2Val(X, threshold = 1, value = 0, below = TRUE)
}
\arguments{
\item{X}{a matrix of vector of values}

\item{threshold}{value above/below which to adjust values. Default is 1.}

\item{value}{a numeric value to set all values below threshold to. Default is 0.}

\item{below}{adjusts values below threshold as default. Adjusts values above threshold if FALSE.}
}
\value{
the adjusted matrix or vector as a data frame
}
\description{
This function sets all values in a vector/matrix that are less/greater than or equal
to a specified threshold equal to a specified value and returns the output as a
data frame.
}
