% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toMatchEpitopes}
\alias{toMatchEpitopes}
\title{List of Epitopes for toMatch}
\format{A data frame with 1 variable:
 \describe{
  \item{id}{id used to identify the epitope recognized}}}
\usage{
data(toMatchEpitopes)
}
\description{
List of Epitopes screend for toMatch
}
\keyword{datasets}
