% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toSearch1Epitopes}
\alias{toSearch1Epitopes}
\title{List of Epitopes for toSearch1}
\format{A data frame with 1 variable:
 \describe{
  \item{id}{id used to identify the epitope recognized}}}
\usage{
data(toSearch1Epitopes)
}
\description{
List of Epitopes screend for toSearch1
}
\keyword{datasets}
