% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toSearch2Epitopes}
\alias{toSearch2Epitopes}
\title{List of Epitopes for toSearch2}
\format{A data frame with 1 variable:
 \describe{
  \item{id}{id used to identify the epitope recognized}}}
\usage{
data(toSearch2Epitopes)
}
\description{
List of Epitopes screend for toSearch2
}
\keyword{datasets}
