% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{wCor.df}
\alias{wCor.df}
\title{Computes weighted correlation between two data frames}
\usage{
wCor.df(
  toSearch,
  samplesToMatch = NULL,
  prevalence = NULL,
  a.par = 0.9,
  d.par = 0.03,
  f.par = 0.75
)
}
\arguments{
\item{toSearch}{a database matrix of z scores for samples to be compared against sampletoMatch or vector of z scores for a single sample}

\item{samplesToMatch}{a dataframe of z scores from a sample to compare against database. Must contain same number of rows as toSearch. Each column represents a single sample}

\item{prevalence}{a vector of epitope prevalence for each epitope id in samplesToMatch and toSearch}

\item{a.par, d.par, f.par}{parameters in weight equation}
}
\value{
weighted correlation data frame
}
\description{
This function computes the weighted correlations between each column of samplesToMatch with each column of toSearch and returns the output
as a data frame. Uses C for computation. Columns containing epitope IDs must be named "id"
}
