% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{wCor.fxn}
\alias{wCor.fxn}
\title{Computes weighted correlation between vector/matrix and vector}
\usage{
wCor.fxn(toSearch, sampleToMatch = NULL, weight = NULL)
}
\arguments{
\item{toSearch}{a database matrix of z scores for samples to be compared against sampletoMatch or vector of z scores for a single sample}

\item{sampleToMatch}{a vector of z scores from a sample to compare against database. Must contain same number of rows as toSearch.}

\item{weight}{a vector of weights. Must have same dimensions as sampleToMatch}
}
\value{
weighted correlation matrix
}
\description{
This function computes the weighted correlation between a vector and each column of a dataframe with the same number or rows and returns the output
as a data frame. Uses C for computation.
}
