% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation.R
\name{z2Hit}
\alias{z2Hit}
\title{generates a matrix describing epitope prescense/absence from z matrix. Prevalence column optional}
\usage{
z2Hit(zDF, t = 3, prev = FALSE, keepID = TRUE)
}
\arguments{
\item{zDF}{a dataframe of Z scores wherer columns represent individuals and rows represent epitopes}

\item{t}{threshold value. Sample will be considered negative for antibodies to epitope for Z-scores less than or equal to t. Corresponding entries in output will be 0.}

\item{prev}{logical. Last column in output will contain prevalence of each eptipe if TRUE}

\item{keepID}{logical. first column will contain epitope ids if TRUE}
}
\value{
a hit matrix with or without optional prevalence column
}
\description{
Takes a matrix of z scores and outputs a binary matrix that describes whether each z score in the
input matrix is above or less than or equal to a user-defined threshold. Each column should represent
a sample and each row should represent an epitope. Option to create a column describing prevalence
within the data set of antibodies to each eptitope based onthe selected threshold.
}
